Add user account management and preference updates
This commit adds in the same user account management as in MahjongSite.
When starting with no user accounts, it will prompt to create one and make
it an admin.  It allows users to update their preferences, using "preferences"
to distinguish it from the tournament "settings".  The preference management
code was moved to its own file, preferences.py.

A setting in mysettings.py called TOURNAMENTNAME is used for the name
of the tournament in email communications.