Changed user editing and added method to inactive players
This commit adds an Inactive flag to the Players table. Only
authenticated users are presented controls for updaing player
info.  Updates to user settings are checked for validity before
commiting them.  If they aren't valid, the field changes class
and is highlighted.  Highlight disappears if the value is
corrected to something valid (and it is commited).

At the end of the player list, an Add Player button allows the creation of
a new player record. New players are created as inactive and must
be activated.

Still to do:
* Give new player a name that always sorts to the end of the list
* Make colors reasonable for default and other preferences
* Make field widths more reasonable
* Enable the "Make inactive", "Reactivate" button