Updated leaderboard to show games played and player status
The leaderboard now has columns for games played and status.
Sorting is based on 3 keys: active/inactive, games played, and score.
This puts inactive players on the bottom and ranks players with the
most games played higher than those with fewer games played.  The
"place" value now correctly accounts for tie scores.

Added a helper utility to load some scores into the database,  It
assumes that there are already some players loaded with IDs 1-12 and
a round 1.