Changed ajax data loading and rendering of tournament tabs
This commit changes the loading and rendering of the tournament tabs
to being on demand.  When the tournament template is loaded, the players
page is shown by default and it is the only one that is loaded.  Clicking
on the tabs loads the other pages (standings, settings, seating).  The
data fetched by renderTemplate is cached in the browser.  It is not
reloaded when the Players column headers are clicked to re-sort the
list of players.

The settings tab is now hidden if not logged in as an admin.

Fixed initial state of showinactive checkbox on Players page.
Changed handler for column header clicks to ensure it did not
respond to clicks on the checkbox for showinactive players (which
is also in the columen header)

Fixed display of tournament name as title of login page.

Commented out a few console.log statements in tournament.js.