Add User managment for admin users
This commit sets up a new tab to manage user accounts.  It can be used
to edit the email address (account name) of the account and set whether
it has admin privilege or not.  It has buttons for adding and deleting
accounts that are non-functioning.

Made some appearance tweaks to see users more clearly, but this still
needs improvement.