Enable Admins to reset passwords of users
This commit adds a Reset Password button to the User management page.
It uses the same mechanism as user-requested password resets to create
a ResetLink and then validate the new password.

Also fixes disappearing edits when re-sorting the users.  The user
attributes are saved in the database, but the re-sort was using a
copy of the user data cached in the browser.  This changes forces it
to reload on every re-sort (assuming a small number of user accounts,
this shouldn't be too long).