Added more error handling to player CSV upload
This update checks for empty rows or rows that look like column
headers in the player CSV file.  After upload it reports the
count of good and bad records found.  Empty values in the
spreadsheet are treated as nulls for the database.  That allows
spreadsheets with no number column filled in to be loaded without
hitting the uniqueness constraint on player number.  The player
type, however, is forced to 0 if null.