Created Penalties table to replace chombos and allow other penalties
This commit alters the database to store penalties in their own
table.  The chombos field in Scores is eliminated.  Penalties is
a daughter table of Scores and allows multiple penalties to be
assessed on a player for a particular game.  The Leaderboard and
Scoreboard are updated to correctly display the total score for
players.

This commit does not have interactive editing of penalties; only
storage and display.  It does include an untested handler for
storing edited penalties for a particular score record.  It includes
some icon images for opening and closing the future penalty editing
section.

It takes a couple of seconds to load the Seating screen.  It's
a more complex SQL query to get the data with the penalty totals
and the page structure is more complex too.  This probably needs
improvement.

Modified loadscores.py helper script to not load chombos.