Add support to store and retrieve unused points for games
Enabled the storage of Scores records for a special player, the
unused points player, to account for riichi bets left over at the
end of a game.  A new player type is used to distinguish the
player record.  When unused points are specified, a new Scores
record will be inserted for the game and it will be associated
with the unused points player.  A new setting,
UNUSEDPOINTSINCREMENT, controls the minimum amount of raw score
points that can be assigned as unused.  All values must be a
multiple of this increment.

This commit only implements storage and retrieval of unused points.
They are not exposed in the GUI yet, and the unused points player is
not hidden from other views of players.

Also:
* Fixed a bug in the User editor where email addresses weren't
  being stored in lowercase.
* Fixed a bug in the login dialog where the next URI wasn't
  specified.