Add interactive editing of unused points for games
The Unused Points entry field appears as soon as any scores are changed in a
game on the Seater tab.  The points are included in the points total for the
game.  Once the points total is 120,000, the entries are saved.  If the unused
points are non-zero, they remain on the Seater display.  Unused points do not
show up in game history displays on the player stats page.