Enhance flexibility in loading players.csv
Added header line detection in players.csv upload.  If the file has a
header, it must be a subset of the column names in the download
players.csv. The columns can be in any order and column names are case
insensitive. At least 3 columns must be provided for every row in
order for it to be accepted One of those columns must be the player
name.  Empty columns are filled with nulls.  Null values inserted in
the Number column do not violate its uniqueness constraint.