Changed Users tab to use Tornado rendering and added Invite managment
Switched rendering of the Users tab from mustache invoked from Javascript
to Tornado's python engine.  This makes it more consistent with other tabs.

Pending invitations to sign up for an account are shown under the list
of current users in the Users tab.  Invitations can be dropped by admins.

Fixed some database constraints so that references to deleted user accounts
wouldn't break other tables.

Added hidden fields to pass the next URI and task name to the reset password
handler.  This is used when managing user accounts where an admin resets
a user's password.  I wanted the next task to be to return to the Users
managment tab.  Unfortunately, jumping to /users.html causes problems.
I'll need to fix that later.