Fixed bug that set country to ABW when changing player type
Changing a player to type = Substitute forced their country code to be
SUB, but when substitue players were set to type regular or Inactive,
it would change their country code to the first country code in the
list.  This commit eliminates the latter kind of country code change.

Added a lookup table for player types instead of a sequential
search through the list of types.