Enable editing of extended tournament attributes
This commit enables creation, editing, and deletion of tournaments in
a single dialog, edittournament.html.  The data is sanity checked before
committing.  Database commits are made as fields change in the browser;
there is no explicit Save button.

Only the tournament owner and admins may edit or delete a tournament.  Users
must be logged in to create a tournament.

The logo and link URL fields don't do anything yet.  The dates are only used
in displaying the tournaments and sorting them by date.