Fix updating country in Tournament Settings and prettify listing
This fixes the problem where picking a different country for the tournament
was not updating the database.  It also changes the layout of the table
of tournament attributes to be prettier.  Messages about successful updates
to attributes are removed. New messages show up for invalid data entries
along with the highlighting of the input fields.