Change creation of new tournaments to happen sooner
This change changes the creation of the Tournaments record to happen
before loading the web page to edit it.  This avoids having to reload
the page on the first change (but now allows tournaments with no name
to be created).  Keyboard focus is forced on the tournament name field
if it's empty.