Store cuts in seating and display them in seating & leaderboard
This change adds a CutName field to the Seating table so each player
table can be assigned to a differnt cut.  The Seating and Standings
pages show the cut names on dividing lines in the listing if any are
assigned when generating the seating for a round.  There is some
slight color and text size variation on the cut nmae separators to
help break-up the listing.

This also fixes the seater to use the CutCount for a round when
determining how many cuts to make.

Leaderboard sorting is changed to sort first by cut name,
then by type of player, and then by score.