Playerstats works across tournaments and no longer uses mustache
Without mustache, the data is mostly formatted by tornado templates.
Th rank histogram is stored in the page and then rendered by d3 calls
on plage load.

This has only been tested on players with a single tournment.  The
framework for using jquery-ui tabs is in place to switch between
multiple tournaments.