Add endpoint to initialize/update countries and use it on first launch
The updateCountries endpoint will check the countries in the database
against those of the countries.csv file.  The comparison is done using
the unique country Code field, not the rowID.  It inserts any new
countries it finds in the file and updates any existing ones that
differ from the file's contents.  By updating without chaning the Code
or Id fields, it should not impact any database references already in
the system.  It reports how many entries were found in the database,
how many were inserted and how may were updated.

When the application is first run on an empty database, instead of
automatically creating the Countries table, it offers to create it
once an administrator accoout is set up.

There's no other link to the updateCountries endpoint, so admins
will just have to manually enter it if updates are needed.  That
should only occur when the countries.csv is changed (when a country
is renamed, or changes its flag, for example).