Add penalties to games in playerstats (correctly)
The playerstats now shows penalties assessed to a player in the games
played section and the total penalties assessed in the tournament summary
(or across all tournaments).  This fixes a bug where multiple penalties
on the same player game+round score would cause the average score and rank
to be miscalculated.