Fix display of players to prevent unaauthorized editing
Turns off display of player type selection when unauthenticated.
Correctly handles POST requests to change player type when unauthenticated.
Enables toggling of DEVELOPERMODE by the presence/absence of a file with
that name in the directory where tomas is launched.  This makes it
possible to switch between developer mode and authentication required
mode without restarting the server.